% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatgpt_addin.R
\name{chatgpt_addin}
\alias{chatgpt_addin}
\title{Assistance ChatGPT Addin}
\usage{
chatgpt_addin()
}
\value{
Lance le gadget Shiny dans le Viewer pane.
}
\description{
Ouvre un gadget Shiny dans le Viewer pane pour poser des questions à l'IA ChatGPT.
}
